common/src/main/java/dto/PointDTO.java
package dto;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

public class PointDTO {
    private String autoId;
    private double lat;
    private double lon;
    private double azimuth;
    private int speed;
    private long time;

    public PointDTO(String autoId,
            double lat,
            double lon,
            double azimuth,
            int speed,
            long time) {
        this.autoId=autoId;
        this.lat=lat;
        this.lon=lon;
        this.azimuth=azimuth;
        this.speed=speed;
        this.time=time;
    }

    public PointDTO(){

    }

    public String getAutoId() {
        return autoId;
    }

    public void setAutoId(String autoId) {
        this.autoId = autoId;
    }

    public double getLat() {
        return lat;
    }

    public void setLat(double lat) {
        this.lat = lat;
    }

    public double getLon() {
        return lon;
    }

    public void setLon(double lon) {
        this.lon = lon;
    }

    public double getAzimuth() {
        return azimuth;
    }

    public void setAzimuth(double azimuth) {
        this.azimuth = azimuth;
    }

    public int getSpeed() {
        return speed;
    }

    public void setSpeed(int speed) {
        this.speed = speed;
    }

    public long getTime() {
        return time;
    }

    public void setTime(long time) {
        this.time = time;
    }

    public String toJson() throws JsonProcessingException {
        ObjectMapper mapper = new ObjectMapper();
        return mapper.writeValueAsString(this);
    }
}
